refactor(condo): DOMA-8629 renamed model for better understanding
