+++ apps/condo/domains/meter/gql.js
@@ -83,8 +83,8 @@ const MeterReadingsImportTask = generateGqlQueries('MeterReadingsImportTask', ME
 const METER_READING_EXPORT_TASK_FIELDS = `{ status format exportedRecordsCount totalRecordsCount file { id originalFilename publicUrl mimetype } meta where sortBy locale timeZone user { id } ${COMMON_FIELDS} }`
 const MeterReadingExportTask = generateGqlQueries('MeterReadingExportTask', METER_READING_EXPORT_TASK_FIELDS)
 
-const CUSTOM_METER_NAME_FIELDS = `{ user { id } meter { id } name ${COMMON_FIELDS} }`
-const CustomMeterName = generateGqlQueries('CustomMeterName', CUSTOM_METER_NAME_FIELDS)
+const METER_USER_DATA_FIELDS = `{ user { id } meter { id } name ${COMMON_FIELDS} }`
+const MeterUserData = generateGqlQueries('MeterUserData', METER_USER_DATA_FIELDS)
 
 /* AUTOGENERATE MARKER <CONST> */
 
@@ -106,7 +106,7 @@ module.exports = {
     REGISTER_METERS_READINGS_MUTATION,
     MeterReadingsImportTask,
     MeterReadingExportTask,
-    CustomMeterName,
+    MeterUserData,
 /* AUTOGENERATE MARKER <EXPORTS> */
 }
