--- apps/condo/domains/meter/schema/CustomMeterName.test.js
@@ -1,285 +1,0 @@
-/**
- * Generated by `createschema meter.CustomMeterName 'user:Relationship:User:CASCADE; meter:Relationship:Meter:CASCADE; name:Text'`
- */
-
-const { faker } = require('@faker-js/faker')
-
-const { makeLoggedInAdminClient, makeClient, UUID_RE, expectValuesOfCommonFields,
-    expectToThrowUniqueConstraintViolationError,
-} = require('@open-condo/keystone/test.utils')
-const {
-    expectToThrowAuthenticationErrorToObj, expectToThrowAuthenticationErrorToObjects,
-    expectToThrowAccessDeniedErrorToObj, expectToThrowAccessDeniedErrorToObjects,
-    expectToThrowGQLError,
-} = require('@open-condo/keystone/test.utils')
-
-const {
-    makeContextWithOrganizationAndIntegrationAsAdmin,
-    createTestBillingProperty,
-    createTestBillingAccount,
-} = require('@condo/domains/billing/utils/testSchema')
-const { COLD_WATER_METER_RESOURCE_ID } = require('@condo/domains/meter/constants/constants')
-const { MeterResource, createTestMeter } = require('@condo/domains/meter/utils/testSchema')
-const { CustomMeterName, createTestCustomMeterName, updateTestCustomMeterName } = require('@condo/domains/meter/utils/testSchema')
-const { makeResidentWithOwnMeter } = require('@condo/domains/meter/utils/testSchema')
-const { createTestOrganization } = require('@condo/domains/organization/utils/testSchema')
-const { createTestProperty } = require('@condo/domains/property/utils/testSchema')
-const {
-    createTestResident,
-    createTestServiceConsumer,
-} = require('@condo/domains/resident/utils/testSchema')
-const { makeClientWithNewRegisteredAndLoggedInUser, makeClientWithSupportUser } = require('@condo/domains/user/utils/testSchema')
-const { makeClientWithResidentUser } = require('@condo/domains/user/utils/testSchema')
-
-
-describe('CustomMeterName', () => {
-    let admin, support, user, anonymous, resident
-    let organization, property, resource, meter
-
-    beforeAll(async () => {
-        admin = await makeLoggedInAdminClient()
-        support = await makeClientWithSupportUser()
-        user = await makeClientWithNewRegisteredAndLoggedInUser()
-        resident = await makeClientWithResidentUser()
-        anonymous = await makeClient();
-
-        [organization] = await createTestOrganization(admin);
-        [resource] = await MeterResource.getAll(admin, { id: COLD_WATER_METER_RESOURCE_ID })
-    })
-    beforeEach(async () => {
-        [property] = await createTestProperty(admin, organization);
-        [meter] = await createTestMeter(admin, organization, property, resource, {})
-    })
-
-    describe('CRUD tests', () => {
-        describe('create', () => {
-            test('admin can', async () => {
-                const [obj, attrs] = await createTestCustomMeterName(admin, admin.user, meter)
-                expectValuesOfCommonFields(obj, attrs, admin)
-            })
-
-            test('support can', async () => {
-                const [obj, attrs] = await createTestCustomMeterName(support, support.user, meter)
-
-                expect(obj.id).toMatch(UUID_RE)
-                expect(obj.dv).toEqual(1)
-                expect(obj.sender).toEqual(attrs.sender)
-                expect(obj.createdBy).toEqual(expect.objectContaining({ id: support.user.id }))
-            })
-
-            test('resident can', async () => {
-                const unitName = faker.random.alphaNumeric(8)
-                const client = await makeClientWithResidentUser()
-                const { context, organization } = await makeContextWithOrganizationAndIntegrationAsAdmin();
-                [property] = await createTestProperty(admin, organization)
-                const [billingProperty] = await createTestBillingProperty(admin, context)
-                const [billingAccount] = await createTestBillingAccount(admin, context, billingProperty)
-                const [resident] = await createTestResident(admin, client.user, property, {
-                    unitName,
-                })
-                await createTestServiceConsumer(admin, resident, organization, {
-                    accountNumber: billingAccount.number,
-                })
-
-                const [meter] = await createTestMeter(admin, organization, property, resource, {
-                    accountNumber: billingAccount.number,
-                    unitName,
-                })
-                const [obj, attrs] = await createTestCustomMeterName(client, resident.user, meter)
-
-                expect(obj.id).toMatch(UUID_RE)
-                expect(obj.dv).toEqual(1)
-                expect(obj.sender).toEqual(attrs.sender)
-                expect(obj.createdBy).toEqual(expect.objectContaining({ id: resident.user.id }))
-                expect(obj.user).toEqual(expect.objectContaining({ id: resident.user.id }))
-                expect(obj.meter).toEqual(expect.objectContaining({ id: meter.id }))
-            })
-
-            test('user can\'t', async () => {
-                await expectToThrowAccessDeniedErrorToObj(async () => {
-                    await createTestCustomMeterName(user, user.user, meter)
-                })
-            })
-
-            test('anonymous can\'t', async () => {
-                await expectToThrowAuthenticationErrorToObj(async () => {
-                    await createTestCustomMeterName(anonymous, user.user, meter)
-                })
-            })
-        })
-
-        describe('update', () => {
-            test('admin can', async () => {
-                const [objCreated] = await createTestCustomMeterName(admin, admin.user, meter)
-
-                const [obj, attrs] = await updateTestCustomMeterName(admin, objCreated.id)
-
-                expect(obj.dv).toEqual(1)
-                expect(obj.sender).toEqual(attrs.sender)
-                expect(obj.v).toEqual(2)
-                expect(obj.updatedBy).toEqual(expect.objectContaining({ id: admin.user.id }))
-            })
-
-
-            test('support can', async () => {
-                const [objCreated] = await createTestCustomMeterName(admin, admin.user, meter)
-
-                const [obj, attrs] = await updateTestCustomMeterName(support, objCreated.id)
-
-                expect(obj.id).toMatch(UUID_RE)
-                expect(obj.dv).toEqual(1)
-                expect(obj.sender).toEqual(attrs.sender)
-                expect(obj.v).toEqual(2)
-                expect(obj.updatedBy).toEqual(expect.objectContaining({ id: support.user.id }))
-            })
-
-            test('resident can update own', async () => {
-                const { residentClient, meter } = await makeResidentWithOwnMeter(admin)
-                const [objCreated] = await createTestCustomMeterName(residentClient, residentClient.user, meter)
-
-                const [obj, attrs] = await updateTestCustomMeterName(residentClient, objCreated.id)
-
-                expect(obj.id).toMatch(UUID_RE)
-                expect(obj.dv).toEqual(1)
-                expect(obj.sender).toEqual(attrs.sender)
-                expect(obj.v).toEqual(2)
-                expect(obj.updatedBy).toEqual(expect.objectContaining({ id: residentClient.user.id }))
-            })
-
-            test('resident can\'t update someone else\'s', async () => {
-                const [objCreated] = await createTestCustomMeterName(admin, admin.user, meter)
-
-                await expectToThrowAccessDeniedErrorToObj(async () => {
-                    await updateTestCustomMeterName(resident, objCreated.id)
-                })
-            })
-
-            test('user can\'t', async () => {
-                const [objCreated] = await createTestCustomMeterName(admin, admin.user, meter)
-                await expectToThrowAccessDeniedErrorToObj(async () => {
-                    await updateTestCustomMeterName(user, objCreated.id)
-                })
-            })
-
-            test('anonymous can\'t', async () => {
-                const admin = await makeLoggedInAdminClient()
-                const [objCreated] = await createTestCustomMeterName(admin, admin.user, meter)
-
-                await expectToThrowAuthenticationErrorToObj(async () => {
-                    await updateTestCustomMeterName(anonymous, objCreated.id)
-                })
-            })
-        })
-
-        describe('hard delete', () => {
-            test('admin can\'t', async () => {
-                const [objCreated] = await createTestCustomMeterName(admin, admin.user, meter)
-
-                await expectToThrowAccessDeniedErrorToObj(async () => {
-                    await CustomMeterName.delete(admin, objCreated.id)
-                })
-            })
-
-            test('user can\'t', async () => {
-                const [objCreated] = await createTestCustomMeterName(admin, admin.user, meter)
-
-                await expectToThrowAccessDeniedErrorToObj(async () => {
-                    await CustomMeterName.delete(user, objCreated.id)
-                })
-            })
-
-            test('anonymous can\'t', async () => {
-                const [objCreated] = await createTestCustomMeterName(admin, admin.user, meter)
-
-                await expectToThrowAccessDeniedErrorToObj(async () => {
-                    await CustomMeterName.delete(anonymous, objCreated.id)
-                })
-            })
-
-            test('resident can\'t', async () => {
-                const [objCreated] = await createTestCustomMeterName(admin, admin.user, meter)
-
-                await expectToThrowAccessDeniedErrorToObj(async () => {
-                    await CustomMeterName.delete(resident, objCreated.id)
-                })
-            })
-        })
-
-        describe('read', () => {
-            test('admin can', async () => {
-                const [obj] = await createTestCustomMeterName(admin, admin.user, meter)
-
-                const objs = await CustomMeterName.getAll(admin, {}, { sortBy: ['updatedAt_DESC'] })
-
-                expect(objs.length).toBeGreaterThanOrEqual(1)
-                expect(objs).toEqual(expect.arrayContaining([
-                    expect.objectContaining({
-                        id: obj.id,
-                    }),
-                ]))
-            })
-
-            test('user can\'t', async () => {
-                await createTestCustomMeterName(admin, admin.user, meter)
-
-                await expectToThrowAccessDeniedErrorToObjects(async () => {
-                    await CustomMeterName.getAll(user, {}, { sortBy: ['updatedAt_DESC'] })
-                })
-            })
-
-            test('resident can read own', async () => {
-                const { residentClient, meter } = await makeResidentWithOwnMeter(admin)
-                const [obj] = await createTestCustomMeterName(residentClient, residentClient.user, meter)
-                const objs = await CustomMeterName.getAll(residentClient, {}, { sortBy: ['updatedAt_DESC'] })
-
-                expect(objs).toHaveLength(1)
-                expect(objs).toEqual(expect.arrayContaining([
-                    expect.objectContaining({
-                        id: obj.id,
-                    }),
-                ]))
-            })
-
-            test('resident can\'t read other resident\'s custom name', async () => {
-                const { residentClient, meter } = await makeResidentWithOwnMeter(admin)
-                await createTestCustomMeterName(residentClient, residentClient.user, meter)
-                const resident2 = await makeClientWithResidentUser()
-
-                const objs = await CustomMeterName.getAll(resident2, {}, { sortBy: ['updatedAt_DESC'] })
-
-                expect(objs).toHaveLength(0)
-            })
-
-            test('anonymous can\'t', async () => {
-                await createTestCustomMeterName(admin, admin.user, meter)
-
-                await expectToThrowAuthenticationErrorToObjects(async () => {
-                    await CustomMeterName.getAll(anonymous, {}, { sortBy: ['updatedAt_DESC'] })
-                })
-            })
-        })
-    })
-
-    describe('Validation tests', () => {
-        test('Should have correct dv field (=== 1)', async () => {
-            await expectToThrowGQLError(
-                async () => await createTestCustomMeterName(admin, admin.user, meter, { dv: 100 }),
-                {
-                    'code': 'BAD_USER_INPUT',
-                    'type': 'DV_VERSION_MISMATCH',
-                    'message': 'Wrong value for data version number',
-                    'mutation': 'createCustomMeterName',
-                    'variable': ['data', 'dv'],
-                },
-            )
-        })
-
-        test('There can only be one custom name for one meter and one user', async () => {
-            await createTestCustomMeterName(admin, admin.user, meter)
-
-            await expectToThrowUniqueConstraintViolationError(async () => {
-                await createTestCustomMeterName(admin, admin.user, meter)
-            }, 'Custom_meter_name_unique_user_meter')
-        })
-    })
-})
