+++ apps/condo/domains/meter/schema/MeterUserData.js
@@ -1,0 +1,64 @@
+/**
+ * Generated by `createschema meter.MeterUserData 'user:Relationship:User:CASCADE; meter:Relationship:Meter:CASCADE; name:Text'`
+ */
+
+const { historical, versioned, uuided, tracked, softDeleted, dvAndSender } = require('@open-condo/keystone/plugins')
+const { GQLListSchema } = require('@open-condo/keystone/schema')
+
+const access = require('@condo/domains/meter/access/MeterUserData')
+const { RESIDENT } = require('@condo/domains/user/constants/common')
+
+
+const MeterUserData = new GQLListSchema('MeterUserData', {
+    schemaDoc: `Custom user data that a user with type ${RESIDENT} can assign to a specific meter device to make it easier to understand what kind of device it is`,
+    plural: 'MeterUserData',
+    fields: {
+
+        user: {
+            schemaDoc: 'User that created their custom meter name',
+            type: 'Relationship',
+            ref: 'User',
+            isRequired: true,
+            knexOptions: { isNotNullable: true }, // Required relationship only!
+            kmigratorOptions: { null: false, on_delete: 'models.CASCADE' },
+        },
+
+        meter: {
+            schemaDoc: 'Meter for which the user created a custom name',
+            type: 'Relationship',
+            ref: 'Meter',
+            isRequired: true,
+            knexOptions: { isNotNullable: true }, // Required relationship only!
+            kmigratorOptions: { null: false, on_delete: 'models.CASCADE' },
+        },
+
+        name: {
+            schemaDoc: 'Custom meter name that a user assigned to a specific meter',
+            type: 'Text',
+            isRequired: true,
+        },
+
+    },
+    kmigratorOptions: {
+        constraints: [
+            {
+                type: 'models.UniqueConstraint',
+                fields: ['user', 'meter'],
+                condition: 'Q(deletedAt__isnull=True)',
+                name: 'Meter_user_data_unique_user_meter',
+            },
+        ],
+    },
+    plugins: [uuided(), versioned(), tracked(), softDeleted(), dvAndSender(), historical()],
+    access: {
+        read: access.canReadMeterUserData,
+        create: access.canManageMeterUserData,
+        update: access.canManageMeterUserData,
+        delete: false,
+        auth: true,
+    },
+})
+
+module.exports = {
+    MeterUserData,
+}
