+++ apps/condo/domains/meter/schema/MeterUserData.test.js
@@ -1,0 +1,285 @@
+/**
+ * Generated by `createschema meter.MeterUserData 'user:Relationship:User:CASCADE; meter:Relationship:Meter:CASCADE; name:Text'`
+ */
+
+const { faker } = require('@faker-js/faker')
+
+const { makeLoggedInAdminClient, makeClient, UUID_RE, expectValuesOfCommonFields,
+    expectToThrowUniqueConstraintViolationError,
+} = require('@open-condo/keystone/test.utils')
+const {
+    expectToThrowAuthenticationErrorToObj, expectToThrowAuthenticationErrorToObjects,
+    expectToThrowAccessDeniedErrorToObj, expectToThrowAccessDeniedErrorToObjects,
+    expectToThrowGQLError,
+} = require('@open-condo/keystone/test.utils')
+
+const {
+    makeContextWithOrganizationAndIntegrationAsAdmin,
+    createTestBillingProperty,
+    createTestBillingAccount,
+} = require('@condo/domains/billing/utils/testSchema')
+const { COLD_WATER_METER_RESOURCE_ID } = require('@condo/domains/meter/constants/constants')
+const { MeterResource, createTestMeter } = require('@condo/domains/meter/utils/testSchema')
+const { MeterUserData, createTestMeterUserData, updateTestMeterUserData } = require('@condo/domains/meter/utils/testSchema')
+const { makeResidentWithOwnMeter } = require('@condo/domains/meter/utils/testSchema')
+const { createTestOrganization } = require('@condo/domains/organization/utils/testSchema')
+const { createTestProperty } = require('@condo/domains/property/utils/testSchema')
+const {
+    createTestResident,
+    createTestServiceConsumer,
+} = require('@condo/domains/resident/utils/testSchema')
+const { makeClientWithNewRegisteredAndLoggedInUser, makeClientWithSupportUser } = require('@condo/domains/user/utils/testSchema')
+const { makeClientWithResidentUser } = require('@condo/domains/user/utils/testSchema')
+
+
+describe('MeterUserData', () => {
+    let admin, support, user, anonymous, resident
+    let organization, property, resource, meter
+
+    beforeAll(async () => {
+        admin = await makeLoggedInAdminClient()
+        support = await makeClientWithSupportUser()
+        user = await makeClientWithNewRegisteredAndLoggedInUser()
+        resident = await makeClientWithResidentUser()
+        anonymous = await makeClient();
+
+        [organization] = await createTestOrganization(admin);
+        [resource] = await MeterResource.getAll(admin, { id: COLD_WATER_METER_RESOURCE_ID })
+    })
+    beforeEach(async () => {
+        [property] = await createTestProperty(admin, organization);
+        [meter] = await createTestMeter(admin, organization, property, resource, {})
+    })
+
+    describe('CRUD tests', () => {
+        describe('create', () => {
+            test('admin can', async () => {
+                const [obj, attrs] = await createTestMeterUserData(admin, admin.user, meter)
+                expectValuesOfCommonFields(obj, attrs, admin)
+            })
+
+            test('support can', async () => {
+                const [obj, attrs] = await createTestMeterUserData(support, support.user, meter)
+
+                expect(obj.id).toMatch(UUID_RE)
+                expect(obj.dv).toEqual(1)
+                expect(obj.sender).toEqual(attrs.sender)
+                expect(obj.createdBy).toEqual(expect.objectContaining({ id: support.user.id }))
+            })
+
+            test('resident can', async () => {
+                const unitName = faker.random.alphaNumeric(8)
+                const client = await makeClientWithResidentUser()
+                const { context, organization } = await makeContextWithOrganizationAndIntegrationAsAdmin();
+                [property] = await createTestProperty(admin, organization)
+                const [billingProperty] = await createTestBillingProperty(admin, context)
+                const [billingAccount] = await createTestBillingAccount(admin, context, billingProperty)
+                const [resident] = await createTestResident(admin, client.user, property, {
+                    unitName,
+                })
+                await createTestServiceConsumer(admin, resident, organization, {
+                    accountNumber: billingAccount.number,
+                })
+
+                const [meter] = await createTestMeter(admin, organization, property, resource, {
+                    accountNumber: billingAccount.number,
+                    unitName,
+                })
+                const [obj, attrs] = await createTestMeterUserData(client, resident.user, meter)
+
+                expect(obj.id).toMatch(UUID_RE)
+                expect(obj.dv).toEqual(1)
+                expect(obj.sender).toEqual(attrs.sender)
+                expect(obj.createdBy).toEqual(expect.objectContaining({ id: resident.user.id }))
+                expect(obj.user).toEqual(expect.objectContaining({ id: resident.user.id }))
+                expect(obj.meter).toEqual(expect.objectContaining({ id: meter.id }))
+            })
+
+            test('user can\'t', async () => {
+                await expectToThrowAccessDeniedErrorToObj(async () => {
+                    await createTestMeterUserData(user, user.user, meter)
+                })
+            })
+
+            test('anonymous can\'t', async () => {
+                await expectToThrowAuthenticationErrorToObj(async () => {
+                    await createTestMeterUserData(anonymous, user.user, meter)
+                })
+            })
+        })
+
+        describe('update', () => {
+            test('admin can', async () => {
+                const [objCreated] = await createTestMeterUserData(admin, admin.user, meter)
+
+                const [obj, attrs] = await updateTestMeterUserData(admin, objCreated.id)
+
+                expect(obj.dv).toEqual(1)
+                expect(obj.sender).toEqual(attrs.sender)
+                expect(obj.v).toEqual(2)
+                expect(obj.updatedBy).toEqual(expect.objectContaining({ id: admin.user.id }))
+            })
+
+
+            test('support can', async () => {
+                const [objCreated] = await createTestMeterUserData(admin, admin.user, meter)
+
+                const [obj, attrs] = await updateTestMeterUserData(support, objCreated.id)
+
+                expect(obj.id).toMatch(UUID_RE)
+                expect(obj.dv).toEqual(1)
+                expect(obj.sender).toEqual(attrs.sender)
+                expect(obj.v).toEqual(2)
+                expect(obj.updatedBy).toEqual(expect.objectContaining({ id: support.user.id }))
+            })
+
+            test('resident can update own', async () => {
+                const { residentClient, meter } = await makeResidentWithOwnMeter(admin)
+                const [objCreated] = await createTestMeterUserData(residentClient, residentClient.user, meter)
+
+                const [obj, attrs] = await updateTestMeterUserData(residentClient, objCreated.id)
+
+                expect(obj.id).toMatch(UUID_RE)
+                expect(obj.dv).toEqual(1)
+                expect(obj.sender).toEqual(attrs.sender)
+                expect(obj.v).toEqual(2)
+                expect(obj.updatedBy).toEqual(expect.objectContaining({ id: residentClient.user.id }))
+            })
+
+            test('resident can\'t update someone else\'s', async () => {
+                const [objCreated] = await createTestMeterUserData(admin, admin.user, meter)
+
+                await expectToThrowAccessDeniedErrorToObj(async () => {
+                    await updateTestMeterUserData(resident, objCreated.id)
+                })
+            })
+
+            test('user can\'t', async () => {
+                const [objCreated] = await createTestMeterUserData(admin, admin.user, meter)
+                await expectToThrowAccessDeniedErrorToObj(async () => {
+                    await updateTestMeterUserData(user, objCreated.id)
+                })
+            })
+
+            test('anonymous can\'t', async () => {
+                const admin = await makeLoggedInAdminClient()
+                const [objCreated] = await createTestMeterUserData(admin, admin.user, meter)
+
+                await expectToThrowAuthenticationErrorToObj(async () => {
+                    await updateTestMeterUserData(anonymous, objCreated.id)
+                })
+            })
+        })
+
+        describe('hard delete', () => {
+            test('admin can\'t', async () => {
+                const [objCreated] = await createTestMeterUserData(admin, admin.user, meter)
+
+                await expectToThrowAccessDeniedErrorToObj(async () => {
+                    await MeterUserData.delete(admin, objCreated.id)
+                })
+            })
+
+            test('user can\'t', async () => {
+                const [objCreated] = await createTestMeterUserData(admin, admin.user, meter)
+
+                await expectToThrowAccessDeniedErrorToObj(async () => {
+                    await MeterUserData.delete(user, objCreated.id)
+                })
+            })
+
+            test('anonymous can\'t', async () => {
+                const [objCreated] = await createTestMeterUserData(admin, admin.user, meter)
+
+                await expectToThrowAccessDeniedErrorToObj(async () => {
+                    await MeterUserData.delete(anonymous, objCreated.id)
+                })
+            })
+
+            test('resident can\'t', async () => {
+                const [objCreated] = await createTestMeterUserData(admin, admin.user, meter)
+
+                await expectToThrowAccessDeniedErrorToObj(async () => {
+                    await MeterUserData.delete(resident, objCreated.id)
+                })
+            })
+        })
+
+        describe('read', () => {
+            test('admin can', async () => {
+                const [obj] = await createTestMeterUserData(admin, admin.user, meter)
+
+                const objs = await MeterUserData.getAll(admin, {}, { sortBy: ['updatedAt_DESC'] })
+
+                expect(objs.length).toBeGreaterThanOrEqual(1)
+                expect(objs).toEqual(expect.arrayContaining([
+                    expect.objectContaining({
+                        id: obj.id,
+                    }),
+                ]))
+            })
+
+            test('user can\'t', async () => {
+                await createTestMeterUserData(admin, admin.user, meter)
+
+                await expectToThrowAccessDeniedErrorToObjects(async () => {
+                    await MeterUserData.getAll(user, {}, { sortBy: ['updatedAt_DESC'] })
+                })
+            })
+
+            test('resident can read own', async () => {
+                const { residentClient, meter } = await makeResidentWithOwnMeter(admin)
+                const [obj] = await createTestMeterUserData(residentClient, residentClient.user, meter)
+                const objs = await MeterUserData.getAll(residentClient, {}, { sortBy: ['updatedAt_DESC'] })
+
+                expect(objs).toHaveLength(1)
+                expect(objs).toEqual(expect.arrayContaining([
+                    expect.objectContaining({
+                        id: obj.id,
+                    }),
+                ]))
+            })
+
+            test('resident can\'t read other resident\'s custom data', async () => {
+                const { residentClient, meter } = await makeResidentWithOwnMeter(admin)
+                await createTestMeterUserData(residentClient, residentClient.user, meter)
+                const resident2 = await makeClientWithResidentUser()
+
+                const objs = await MeterUserData.getAll(resident2, {}, { sortBy: ['updatedAt_DESC'] })
+
+                expect(objs).toHaveLength(0)
+            })
+
+            test('anonymous can\'t', async () => {
+                await createTestMeterUserData(admin, admin.user, meter)
+
+                await expectToThrowAuthenticationErrorToObjects(async () => {
+                    await MeterUserData.getAll(anonymous, {}, { sortBy: ['updatedAt_DESC'] })
+                })
+            })
+        })
+    })
+
+    describe('Validation tests', () => {
+        test('Should have correct dv field (=== 1)', async () => {
+            await expectToThrowGQLError(
+                async () => await createTestMeterUserData(admin, admin.user, meter, { dv: 100 }),
+                {
+                    'code': 'BAD_USER_INPUT',
+                    'type': 'DV_VERSION_MISMATCH',
+                    'message': 'Wrong value for data version number',
+                    'mutation': 'createMeterUserData',
+                    'variable': ['data', 'dv'],
+                },
+            )
+        })
+
+        test('There can only be one custom data for one meter and one user', async () => {
+            await createTestMeterUserData(admin, admin.user, meter)
+
+            await expectToThrowUniqueConstraintViolationError(async () => {
+                await createTestMeterUserData(admin, admin.user, meter)
+            }, 'Meter_user_data_unique_user_meter')
+        })
+    })
+})
