--- apps/condo/domains/meter/utils/clientSchema/CustomMeterName.ts
@@ -1,30 +1,0 @@
-/**
- * Generated by `createschema meter.CustomMeterName 'user:Relationship:User:CASCADE; meter:Relationship:Meter:CASCADE; name:Text'`
- */
-
-import {
-    CustomMeterName,
-    CustomMeterNameCreateInput,
-    CustomMeterNameUpdateInput,
-    QueryAllCustomMeterNamesArgs,
-} from '@app/condo/schema'
-
-import { generateReactHooks } from '@open-condo/codegen/generate.hooks'
-
-import { CustomMeterName as CustomMeterNameGQL } from '@condo/domains/meter/gql'
-
-const {
-    useObject,
-    useObjects,
-    useCreate,
-    useUpdate,
-    useSoftDelete,
-} = generateReactHooks<CustomMeterName, CustomMeterNameCreateInput, CustomMeterNameUpdateInput, QueryAllCustomMeterNamesArgs>(CustomMeterNameGQL)
-
-export {
-    useObject,
-    useObjects,
-    useCreate,
-    useUpdate,
-    useSoftDelete,
-}
