+++ apps/condo/domains/meter/utils/clientSchema/MeterUserData.ts
@@ -1,0 +1,30 @@
+/**
+ * Generated by `createschema meter.MeterUserData 'user:Relationship:User:CASCADE; meter:Relationship:Meter:CASCADE; name:Text'`
+ */
+
+import {
+    MeterUserData,
+    MeterUserDataCreateInput,
+    MeterUserDataUpdateInput,
+    QueryAllMeterUserDataArgs,
+} from '@app/condo/schema'
+
+import { generateReactHooks } from '@open-condo/codegen/generate.hooks'
+
+import { MeterUserData as MeterUserDataGQL } from '@condo/domains/meter/gql'
+
+const {
+    useObject,
+    useObjects,
+    useCreate,
+    useUpdate,
+    useSoftDelete,
+} = generateReactHooks<MeterUserData, MeterUserDataCreateInput, MeterUserDataUpdateInput, QueryAllMeterUserDataArgs>(MeterUserDataGQL)
+
+export {
+    useObject,
+    useObjects,
+    useCreate,
+    useUpdate,
+    useSoftDelete,
+}
