+++ apps/condo/domains/meter/utils/clientSchema/index.ts
@@ -25,7 +25,7 @@ export * as PropertyMeter from './PropertyMeter'
 export * as PropertyMeterReading from './PropertyMeterReading'
 export * as MeterReportingPeriod from './MeterReportingPeriod'
 export * as MeterResourceOwner from './MeterResourceOwner'
-export * as CustomMeterName from './CustomMeterName'
+export * as MeterUserData from './MeterUserData'
 /* AUTOGENERATE MARKER <IMPORT-EXPORT> */
